testbench: share skeleton file preservation in command tests
Extract the repeated purge/sync skeleton command test preservation logic into a shared PreservesSkeletonFiles trait.

Both command tests now snapshot the exact skeleton files they mutate, restore existing files with their parent directories, and delete files that did not exist before the test. This keeps the shared Testbench runtime skeleton stable across package-mode tests without duplicating file, directory, and symlink cleanup logic in each test class.
